autoinstrumentation/python/requirements.txt
helios-opentelemetry-sdk=v1.0.66